+++ lib/src/variant/musketeer.dart
@@ -1,7 +1,8 @@
-import 'package:bishop/bishop.dart';
+part of 'variant.dart';
 
 class Musketeer {
-  static const String defaultFen = '8/rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR/8 w KQkq - 0 1';
+  static const String defaultFen =
+      '8/rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR/8 w KQkq - 0 1';
   static PieceType leopard() => PieceType.fromBetza('F2N');
   static PieceType hawk() => PieceType.fromBetza('ADGH');
   static PieceType unicorn() => PieceType.fromBetza('NC');